fix: allow using multiple `$onAction`, **ignore returned value**
Fix #893

This change removes the ability to return `false` to avoid an action to
throw because differently from `onErrorCaptured` from Vue 3, errors are
not propagated here and this behavior could lead to actions not throwing
inside components, which is is rather error-prone. This behavior can
still be achieved with a plugin that override every action in a store.
It also removes the ability to override the returned value with `after`,
which was error prone for the same arguments. Fortunately, these were
**undocumented features**, and should not be used. Revisitting the old
implementation could be revisited but should go through an RFC to
discuss different possibilities (e.g. should the returned value inside
`after` be passed to any other `after`?) to find the most one that makes
more sense for a default.